backend/requirements.txt
alembic==1.13.3
amqp==5.3.1
bcrypt==4.2.0
billiard==4.2.1
blinker==1.8.2
cachelib==0.13.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
flask-paginate==2024.4.12
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
kombu==5.5.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.2
msgspec==0.18.6
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
prompt_toolkit==3.0.50
PyHTML==1.3.2
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==78.1.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
alembic==1.13.3
amqp==5.3.1
aniso8601==10.0.0
bcrypt==4.2.0
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
cachelib==0.13.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cssselect2==0.8.0
cycler==0.12.1
dotenv==0.9.9
Flask==3.0.3
Flask-Caching==2.3.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
flask-paginate==2024.4.12
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
kombu==5.5.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.2
msgspec==0.18.6
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
prompt_toolkit==3.0.50
pycparser==2.22
pydyf==0.11.0
PyHTML==1.3.2
PyJWT==2.10.1
pyparsing==3.2.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.0
webencodings==0.5.1
Werkzeug==3.0.4
zopfli==0.2.3.post1
